Support conversational state in agent router
