--- agentic_router/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the 'agentic_router' directory a Python package.
