--- agentic_router/nodes/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the 'nodes' directory a Python package.
