+++ pyproject.toml
@@ -1,0 +1,38 @@
+[project]
+name = "agentic-router"
+version = "0.1.0"
+description = "LangGraph CLI compatible agent routing workflow."
+readme = "README.md"
+requires-python = ">=3.9"
+dependencies = [
+    "httpx>=0.27.0",
+    "langchain-core>=0.2.0",
+    "langgraph>=0.2.6",
+    "langgraph-cli[inmem]>=0.2.8",
+    "pydantic>=2.0.0",
+    "PyYAML>=6.0",
+]
+
+[project.optional-dependencies]
+dev = [
+    "mypy>=1.11.0",
+    "ruff>=0.6.0",
+    "pytest>=8.0.0",
+]
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.ruff]
+line-length = 100
+
+[tool.ruff.lint]
+select = ["E", "F", "I"]
+ignore = ["E501"]
+
+[tool.ruff.lint.isort]
+known-first-party = ["agentic_router"]
