+++ src/agentic_router/__init__.py
@@ -1,0 +1,5 @@
+"""Agentic Router LangGraph package."""
+
+from .graph import graph
+
+__all__ = ["graph"]
