+++ src/agentic_router/nodes/__init__.py
@@ -1,0 +1,1 @@
+"""Node implementations used by the Agentic Router graph."""
